+++ liquid/golden/range_objects.py
@@ -5,17 +5,15 @@ from liquid.golden.case import Case
 cases = [
     Case(
         description="start is not a number",
-        template=r"{{ (start..end) }}",
+        template=r"{{ (start..end) | join: '#' }}",
         globals={"start": "foo", "end": 5},
-        expect="",
-        error=True,
+        expect="0#1#2#3#4#5",
     ),
     Case(
         description="end is not a number",
-        template=r"{{ (start..end) }}",
+        template=r"{{ (start..end) | join: '#' }}",
         globals={"start": "1", "end": "foo"},
         expect="",
-        error=True,
     ),
     Case(
         description="end is less than start",