+++ tests/test_malformed.py
@@ -188,18 +188,6 @@ class MalformedTemplateTestCase(TestCase):
                 ),
             ),
             Case(
-                description="range start does not cast to int",
-                template="{% for x in (tag..4) %}{{ x }}{% endfor %}",
-                expect_exception=LiquidTypeError,
-                expect_msg=("expected an int or float, found 'goodbye', on line 1"),
-            ),
-            Case(
-                description="range end does not cast to int",
-                template="{% for x in (1..tag) %}{{ x }}{% endfor %}",
-                expect_exception=LiquidTypeError,
-                expect_msg=("expected an int or float, found 'goodbye', on line 1"),
-            ),
-            Case(
                 description="missing equal in assignment tag",
                 template="{% assign x 5 %}",
                 expect_exception=LiquidSyntaxError,